Delete redundant parser write code
