+++ binrw_derive/src/codegen/has_magic.rs
@@ -1,5 +1,5 @@
 use super::sanitization::HAS_MAGIC;
-use crate::parser::read::Input;
+use crate::parser::Input;
 use proc_macro2::TokenStream;
 use quote::quote;
