+++ binrw_derive/src/codegen/read_options/map.rs
@@ -1,6 +1,6 @@
 #[allow(clippy::wildcard_imports)]
 use crate::codegen::sanitization::*;
-use crate::parser::read::Input;
+use crate::parser::Input;
 use proc_macro2::{Span, TokenStream};
 use quote::{quote, quote_spanned};
 use syn::spanned::Spanned;
