+++ binrw_derive/src/codegen/write_options/enum.rs
@@ -1,4 +1,4 @@
-use crate::parser::read::{Enum, EnumVariant, Input, UnitEnumField, UnitOnlyEnum};
+use crate::parser::{Enum, EnumVariant, Input, UnitEnumField, UnitOnlyEnum};
 use proc_macro2::{Ident, TokenStream};
 use quote::quote;
