+++ binrw_derive/src/codegen/write_options/struct.rs
@@ -1,4 +1,4 @@
-use crate::parser::read::{Input, Struct};
+use crate::parser::{Input, Struct};
 use proc_macro2::TokenStream;
 use quote::quote;
 use syn::Ident;
