+++ binrw_derive/src/parser/types/assert.rs
@@ -1,4 +1,4 @@
-use crate::parser::{read::attrs, KeywordToken};
+use crate::parser::{attrs, KeywordToken};
 use core::convert::TryFrom;
 use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
