+++ binrw_derive/src/parser/types/read_mode.rs
@@ -1,4 +1,4 @@
-use crate::parser::{read::attrs, KeywordToken, TrySet};
+use crate::parser::{attrs, KeywordToken, TrySet};
 use proc_macro2::TokenStream;
 use quote::ToTokens;
